% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculator_functions.R
\name{calculate_defacto_eaf_matrix}
\alias{calculate_defacto_eaf_matrix}
\title{Calculate De Facto EAF Matrix}
\usage{
calculate_defacto_eaf_matrix(cohort_genomes)
}
\arguments{
\item{cohort_genomes}{A list of matrices where each matrix represents the genomes of a cohort.
Each matrix has individuals as rows and SNPs as columns.}
}
\value{
A matrix where each column represents the EAF for each cohort.
}
\description{
This function calculates the de facto Effect Allele Frequency (EAF) matrix
for a given list of cohort genomes.
}
