% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculator_functions.R
\name{calculate_defacto_sample_size_matrix}
\alias{calculate_defacto_sample_size_matrix}
\title{Calculate De Facto Sample Size Matrix}
\usage{
calculate_defacto_sample_size_matrix(cohort_genomes)
}
\arguments{
\item{cohort_genomes}{A list of matrices representing the genomes of different cohorts.}
}
\value{
A matrix with the number of participants for each cohort.
}
\description{
This function computes the sample size matrix for a given list of cohort genomes.
}
