% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculator_functions.R
\name{calculate_estimated_heritable_correlation_matrix}
\alias{calculate_estimated_heritable_correlation_matrix}
\title{Calculate Estimated Heritable Correlation Matrix}
\usage{
calculate_estimated_heritable_correlation_matrix(summary_stats_matrix)
}
\arguments{
\item{summary_stats_matrix}{A list of matrices containing summary statistics for each phenotype.}
}
\value{
A matrix representing the estimated heritable correlation between phenotypes.
}
\description{
This function estimates the heritable correlation matrix based on summary statistics.
}
