% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculator_functions.R
\name{calculate_genetic_correlation}
\alias{calculate_genetic_correlation}
\title{Calculate Genetic Correlation Matrix}
\usage{
calculate_genetic_correlation(efs_matrix_template, defacto_eaf_matrix)
}
\arguments{
\item{efs_matrix_template}{A matrix of effect sizes with phenotypes as rows and SNPs as columns.}

\item{defacto_eaf_matrix}{A matrix representing the effect allele frequencies across cohorts.}
}
\value{
A genetic correlation matrix.
}
\description{
This function calculates the genetic correlation matrix using the effect size matrix
and the de facto EAF matrix.
}
