% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculator_functions.R
\name{calculate_heritable_phenotype_matrix}
\alias{calculate_heritable_phenotype_matrix}
\title{Calculate Heritable Phenotype Matrix}
\usage{
calculate_heritable_phenotype_matrix(cohort_genomes, efs_matrix)
}
\arguments{
\item{cohort_genomes}{A list of matrices where each matrix represents the genomes of a cohort.}

\item{efs_matrix}{A matrix of effect sizes with phenotypes as rows and SNPs as columns.}
}
\value{
A list of matrices, each representing heritable phenotypes for a cohort.
}
\description{
This function computes the heritable phenotype matrix based on cohort genomes
and the effect size (EFS) matrix.
}
