% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculator_functions.R
\name{calculate_phenotype_matrix}
\alias{calculate_phenotype_matrix}
\title{Calculate Combined Phenotype Matrix}
\usage{
calculate_phenotype_matrix(
  heritable_phenotype_matrix,
  nonheritable_phenotype_matrix
)
}
\arguments{
\item{heritable_phenotype_matrix}{A list of matrices representing heritable phenotypes.}

\item{nonheritable_phenotype_matrix}{A list of matrices representing non-heritable phenotypes.}
}
\value{
A list of matrices with combined phenotypes for each cohort.
}
\description{
This function combines heritable and non-heritable phenotype matrices
to generate the overall phenotype matrix for each cohort.
}
