% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculator_functions.R
\name{calculate_summary_stats_matrix}
\alias{calculate_summary_stats_matrix}
\title{Calculate Summary Statistics Matrix}
\usage{
calculate_summary_stats_matrix(cohort_genomes, cohort_phenotypes)
}
\arguments{
\item{cohort_genomes}{A list of matrices representing the genomes of different cohorts.}

\item{cohort_phenotypes}{A list of matrices representing the phenotypes of different cohorts.}
}
\value{
A list of matrices containing summary statistics for each phenotype.
}
\description{
This function calculates the summary statistics (beta, standard error, p-value, EAF)
for each SNP based on the cohort genomes and phenotypes.
}
