% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creator_functions.R
\name{create_cohort_makeup_matrix}
\alias{create_cohort_makeup_matrix}
\title{Create Cohort Makeup Matrix}
\usage{
create_cohort_makeup_matrix(
  n_phenotype = 3,
  n_participant = 1000,
  unique_cohorts = TRUE
)
}
\arguments{
\item{n_phenotype}{Integer specifying the number of phenotypes. Default is 3.}

\item{n_participant}{Numeric value or vector indicating the number of participants. Default is 1000.}

\item{unique_cohorts}{Logical indicating whether participants are unique to each cohort. Default is TRUE.}
}
\value{
A matrix representing the cohort makeup with the number of participants for each cohort and their intersections.
}
\description{
This function generates a cohort makeup matrix defining the number of participants
for each phenotype and their intersections.
}
