% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creator_functions.R
\name{create_eaf_matrix}
\alias{create_eaf_matrix}
\title{Create Effect Allele Frequency (EAF) Matrix}
\usage{
create_eaf_matrix(
  n_phenotype = 3,
  uniform_eaf = 0.4,
  n_variant_pleiotropic = 10,
  n_variant_nonpleiotropic = c(10, 10, 10),
  n_variant_null = 960
)
}
\arguments{
\item{n_phenotype}{Integer specifying the number of phenotypes. Default is 3.}

\item{uniform_eaf}{Numeric value representing the uniform effect allele frequency. Default is 0.4.}

\item{n_variant_pleiotropic}{Number of pleiotropic variants. Default is 10.}

\item{n_variant_nonpleiotropic}{Vector specifying the number of non-pleiotropic variants per phenotype. Default is c(10, 10, 10).}

\item{n_variant_null}{Number of null variants. Default is 960.}
}
\value{
A matrix representing effect allele frequencies for variants across cohorts.
}
\description{
This function generates an EAF matrix for different variants across multiple cohorts.
}
