% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creator_functions.R
\name{create_efs_matrix}
\alias{create_efs_matrix}
\title{Create Effect Size (EFS) Matrix}
\usage{
create_efs_matrix(
  efs_matrix_template,
  defacto_eaf_matrix,
  heritable_correlation_matrix
)
}
\arguments{
\item{efs_matrix_template}{A matrix template representing initial effect sizes.}

\item{defacto_eaf_matrix}{A matrix representing the effect allele frequencies.}

\item{heritable_correlation_matrix}{A matrix specifying the target heritable correlation
between phenotypes, used to guide the optimization process.}
}
\value{
A matrix representing optimized effect sizes for each variant and phenotype.
}
\description{
This function generates the effect size (EFS) matrix by optimizing parameters
to minimize the loss function based on the genetic correlation. The optimization
aims to align the simulated genetic correlation with the provided heritable correlation matrix.
}
