% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creator_functions.R
\name{create_efs_matrix_template}
\alias{create_efs_matrix_template}
\title{Create Effect Size (EFS) Matrix Template}
\usage{
create_efs_matrix_template(
  n_phenotype = 3,
  n_variant_pleiotropic = 10,
  n_variant_nonpleiotropic = c(10, 10, 10),
  n_variant_null = 960,
  crosstrait_heterogeneity = TRUE,
  withintrait_heterogeneity = TRUE,
  random_crosstrait_heterogeneity = FALSE,
  random_withintrait_heterogeneity = FALSE
)
}
\arguments{
\item{n_phenotype}{Integer specifying the number of phenotypes. Default is 3.}

\item{n_variant_pleiotropic}{Number of pleiotropic variants. Default is 10.}

\item{n_variant_nonpleiotropic}{Vector specifying the number of non-pleiotropic variants per phenotype. Default is c(10, 10, 10).}

\item{n_variant_null}{Number of null variants. Default is 960.}

\item{crosstrait_heterogeneity}{Logical indicating if cross-trait heterogeneity should be applied. Default is TRUE.}

\item{withintrait_heterogeneity}{Logical indicating if within-trait heterogeneity should be applied. Default is TRUE.}

\item{random_crosstrait_heterogeneity}{Logical indicating if cross-trait heterogeneity should be random. Default is FALSE.}

\item{random_withintrait_heterogeneity}{Logical indicating if within-trait heterogeneity should be random. Default is FALSE.}
}
\value{
A matrix template representing the effect sizes for each variant and phenotype.
}
\description{
This function generates a template for the effect size (EFS) matrix,
specifying the effect of each variant on different phenotypes.
}
