% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creator_functions.R
\name{create_heritable_correlation_matrix}
\alias{create_heritable_correlation_matrix}
\title{Create Heritable Correlation Matrix}
\usage{
create_heritable_correlation_matrix(
  n_phenotype = 3,
  uniform_heritable_correlation = 0.4
)
}
\arguments{
\item{n_phenotype}{Integer specifying the number of phenotypes. Default is 3.}

\item{uniform_heritable_correlation}{Numeric value representing the uniform correlation between phenotypes. Default is 0.4.}
}
\value{
A square matrix representing the heritable correlation between phenotypes.
}
\description{
This function generates a heritable correlation matrix with uniform correlation values
for off-diagonal elements and 1 for diagonal elements.
}
