% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creator_functions.R
\name{create_nonheritable_correlation_matrix}
\alias{create_nonheritable_correlation_matrix}
\title{Create Non-Heritable Correlation Matrix}
\usage{
create_nonheritable_correlation_matrix(
  n_phenotype,
  uniform_nonheritable_correlation = 0.2
)
}
\arguments{
\item{n_phenotype}{Integer specifying the number of phenotypes.}

\item{uniform_nonheritable_correlation}{Numeric value representing the uniform correlation between non-heritable phenotypes. Default is 0.2.}
}
\value{
A square matrix representing the non-heritable correlation between phenotypes.
}
\description{
This function generates a non-heritable correlation matrix with uniform correlation values
for off-diagonal elements and 1 for diagonal elements.
}
