% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pleio.R
\docType{class}
\name{pleio-class}
\alias{pleio-class}
\title{Pleiotropic Simulation Class (pleio)}
\description{
Defines the S4 class 'pleio' for storing pleiotropic simulation results,
including genotype data, phenotype data, and summary statistics.
}
\section{Slots}{

\describe{
\item{\code{n_phenotype}}{Number of phenotypes simulated.}

\item{\code{n_variant_pleiotropic}}{Number of pleiotropic variants.}

\item{\code{n_variant_nonpleiotropic}}{Number of non-pleiotropic variants per phenotype.}

\item{\code{n_variant_null}}{Number of null variants.}

\item{\code{heritability}}{Vector of heritability values for each phenotype.}

\item{\code{unique_cohorts}}{Logical indicating if participants are unique across cohorts.}

\item{\code{crosstrait_heterogeneity}}{Logical indicating cross-trait heterogeneity.}

\item{\code{withintrait_heterogeneity}}{Logical indicating within-trait heterogeneity.}

\item{\code{random_crosstrait_heterogeneity}}{Logical for random cross-trait heterogeneity.}

\item{\code{random_withintrait_heterogeneity}}{Logical for random within-trait heterogeneity.}

\item{\code{heritable_correlation_matrix}}{Matrix of heritable correlation values.}

\item{\code{nonheritable_correlation_matrix}}{Matrix of non-heritable correlation values.}

\item{\code{cohort_makeup_matrix}}{Matrix defining the cohort structure.}

\item{\code{eaf_matrix}}{Matrix of effect allele frequencies.}

\item{\code{efs_matrix_template}}{Template matrix for effect size simulation.}

\item{\code{efs_matrix}}{Matrix of effect sizes.}

\item{\code{heritable_phenotype_matrix}}{List of heritable phenotype data.}

\item{\code{nonheritable_phenotype_matrix}}{List of non-heritable phenotype data.}

\item{\code{defacto_eaf_matrix}}{Matrix of de facto effect allele frequencies.}

\item{\code{defacto_sample_size_matrix}}{Matrix of sample sizes for each cohort.}

\item{\code{estimated_heritable_correlation_matrix}}{Estimated matrix of heritable correlations.}

\item{\code{cohort_genomes}}{List of genotype data for each cohort.}

\item{\code{cohort_phenotypes}}{List of phenotype data for each cohort.}

\item{\code{summary_stats_matrix}}{List of summary statistics for each phenotype.}
}}

