% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pleiosim.R
\name{pleiosim}
\alias{pleiosim}
\title{Pleiotropic Simulation (pleiosim)}
\usage{
pleiosim(
  n_phenotype = 3,
  heritable_correlation = 0.4,
  nonheritable_correlation = 0.2,
  n_participant = 10000,
  eaf = 0.4,
  n_variant_pleiotropic = 10,
  n_variant_nonpleiotropic = c(10, 10, 10),
  n_variant_null = 960,
  heritability = c(0.1, 0.2, 0.3),
  unique_cohorts = TRUE,
  crosstrait_heterogeneity = TRUE,
  withintrait_heterogeneity = TRUE,
  random_crosstrait_heterogeneity = FALSE,
  random_withintrait_heterogeneity = FALSE,
  customized_heritable_correlation_matrix = NULL,
  customized_nonheritable_correlation_matrix = NULL,
  customized_cohort_makeup_matrix = NULL,
  customized_eaf_matrix = NULL,
  customized_efs_matrix_template = NULL,
  customized_efs_matrix = NULL
)
}
\arguments{
\item{n_phenotype}{Number of phenotypes. Default is 3.}

\item{heritable_correlation}{Uniform heritable correlation value. Default is 0.4.}

\item{nonheritable_correlation}{Uniform non-heritable correlation value. Default is 0.2.}

\item{n_participant}{Number of participants. Default is 10000.}

\item{eaf}{Effect allele frequency. Default is 0.4.}

\item{n_variant_pleiotropic}{Number of pleiotropic variants. Default is 10.}

\item{n_variant_nonpleiotropic}{Number of non-pleiotropic variants per phenotype. Default is c(10, 10, 10).}

\item{n_variant_null}{Number of null variants. Default is 960.}

\item{heritability}{Vector of heritability values for each phenotype. Default is c(0.1, 0.2, 0.3).}

\item{unique_cohorts}{Logical indicating whether participants are unique across cohorts. Default is TRUE.}

\item{crosstrait_heterogeneity}{Logical for cross-trait heterogeneity. Default is TRUE.}

\item{withintrait_heterogeneity}{Logical for within-trait heterogeneity. Default is TRUE.}

\item{random_crosstrait_heterogeneity}{Logical for random cross-trait heterogeneity. Default is FALSE.}

\item{random_withintrait_heterogeneity}{Logical for random within-trait heterogeneity. Default is FALSE.}

\item{customized_heritable_correlation_matrix}{Optional custom heritable correlation matrix.}

\item{customized_nonheritable_correlation_matrix}{Optional custom non-heritable correlation matrix.}

\item{customized_cohort_makeup_matrix}{Optional custom cohort makeup matrix. Ensure correct row/column names.}

\item{customized_eaf_matrix}{Optional custom EAF matrix. Ensure correct structure with appropriate row/column names.}

\item{customized_efs_matrix_template}{Optional custom EFS matrix template. Follow given data structure.}

\item{customized_efs_matrix}{Optional custom EFS matrix. Ensure consistency with matrix structure.}
}
\value{
A "pleio" class object containing simulated genetic, phenotype, and summary statistics data.
}
\description{
This function performs a comprehensive simulation of pleiotropic genetic data,
including genotype generation, phenotype calculation, and summary statistics.
}
\details{
For `customized_cohort_makeup_matrix`, `customized_eaf_matrix`, `customized_efs_matrix_template`,
and `customized_efs_matrix`, users can provide custom matrices by specifying the data directly.
Ensure the provided matrices follow the given data structure, including proper row and column names.
}
\examples{
pleio_object = pleiosim(
  n_phenotype = 3,
  heritable_correlation = 0.4,
  nonheritable_correlation = 0.2,
  n_participant = 10000,
  eaf = 0.4,
  n_variant_pleiotropic = 10,
  n_variant_nonpleiotropic = c(10, 10, 10),
  n_variant_null = 960,
  heritability = c(0.1, 0.2, 0.3),
  unique_cohorts = TRUE,
  crosstrait_heterogeneity = TRUE,
  withintrait_heterogeneity = TRUE,
  random_crosstrait_heterogeneity = FALSE,
  random_withintrait_heterogeneity = FALSE,
  customized_heritable_correlation_matrix = NULL,
  customized_nonheritable_correlation_matrix = NULL,
  customized_cohort_makeup_matrix = NULL,
  customized_eaf_matrix = NULL,
  customized_efs_matrix_template = NULL,
  customized_efs_matrix = NULL
)

summary(pleio_object)

}
