% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pleio.R
\name{show,pleio-method}
\alias{show,pleio-method}
\title{Display Method for pleio Class}
\usage{
\S4method{show}{pleio}(object)
}
\arguments{
\item{object}{An object of class 'pleio'.}
}
\description{
Custom display method for objects of class 'pleio' to summarize key simulation parameters.
}
