% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator_functions.R
\name{simulate_genomes}
\alias{simulate_genomes}
\title{Simulate Genomes}
\usage{
simulate_genomes(n_participant, eaf)
}
\arguments{
\item{n_participant}{Integer specifying the number of participants.}

\item{eaf}{Numeric vector representing the effect allele frequencies for each SNP.}
}
\value{
A matrix of simulated genotypes with individuals as rows and SNPs as columns.
}
\description{
This function simulates genotype data for a cohort based on the specified
effect allele frequency (EAF). It generates genotypes with probabilities
corresponding to homozygous (2), heterozygous (1), and wild-type (0) states.
}
