% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator_functions.R
\name{simulate_nonheritable_phenotype_matrix}
\alias{simulate_nonheritable_phenotype_matrix}
\title{Simulate Non-Heritable Phenotype Matrix}
\usage{
simulate_nonheritable_phenotype_matrix(
  heritable_phenotype_matrix,
  nonheritable_correlation_matrix,
  defacto_sample_size_matrix,
  heritability
)
}
\arguments{
\item{heritable_phenotype_matrix}{A list of matrices representing the heritable component of phenotypes.}

\item{nonheritable_correlation_matrix}{A matrix specifying the correlation between non-heritable phenotypes.}

\item{defacto_sample_size_matrix}{A matrix indicating the sample size for each cohort.}

\item{heritability}{A numeric vector representing the heritability for each phenotype.}
}
\value{
A list of matrices where each matrix represents non-heritable phenotypes for a cohort.
}
\description{
This function simulates non-heritable phenotypes based on the specified
non-heritable correlation matrix, sample sizes, and heritability values. It accounts
for environmental variance to generate realistic non-genetic influences on phenotypes.
}
